+++ yash-semantics/src/command_impl/item.rs
@@ -26,6 +26,7 @@ use yash_env::semantics::ExitStatus;
 use yash_env::semantics::Result;
 use yash_env::Env;
 use yash_syntax::source::Location;
+use yash_syntax::source::Span;
 use yash_syntax::syntax;
 use yash_syntax::syntax::AndOrList;
 
@@ -58,7 +59,7 @@ impl Command for syntax::Item {
     }
 }
 
-async fn execute_async(env: &mut Env, and_or: &Rc<AndOrList>, async_flag: &Location) -> Result {
+async fn execute_async(env: &mut Env, and_or: &Rc<AndOrList>, async_flag: &Span) -> Result {
     let and_or = Rc::clone(and_or);
     let result = env
         .start_subshell(|env| Box::pin(async move { and_or.execute(env).await }))
@@ -70,11 +71,15 @@ async fn execute_async(env: &mut Env, and_or: &Rc<AndOrList>, async_flag: &Locat
             Continue(())
         }
         Err(errno) => {
+            let async_flag = Location {
+                code: async_flag.code.clone(),
+                index: async_flag.range.start,
+            };
             print_error(
                 env,
                 "cannot start a subshell to run an asynchronous command".into(),
                 errno.desc().into(),
-                async_flag,
+                &async_flag,
             )
             .await;
 
@@ -121,7 +126,7 @@ mod tests {
         let and_or: syntax::AndOrList = "return -n 42".parse().unwrap();
         let item = syntax::Item {
             and_or: Rc::new(and_or),
-            async_flag: Some(Location::dummy("")),
+            async_flag: Some(Span::dummy("")),
         };
         let result = executor.run_until(item.execute(&mut env));
         assert_eq!(result, Continue(()));
@@ -140,7 +145,7 @@ mod tests {
         let and_or: syntax::AndOrList = "echo foo".parse().unwrap();
         let item = syntax::Item {
             and_or: Rc::new(and_or),
-            async_flag: Some(Location::dummy("")),
+            async_flag: Some(Span::dummy("")),
         };
 
         executor
@@ -170,7 +175,7 @@ mod tests {
         let and_or: syntax::AndOrList = "return -n 42".parse().unwrap();
         let item = syntax::Item {
             and_or: Rc::new(and_or),
-            async_flag: Some(Location::dummy("")),
+            async_flag: Some(Span::dummy("")),
         };
         executor.run_until(item.execute(&mut env));
 
@@ -188,7 +193,7 @@ mod tests {
         let and_or: syntax::AndOrList = "return -n 42".parse().unwrap();
         let item = syntax::Item {
             and_or: Rc::new(and_or),
-            async_flag: Some(Location::dummy("X")),
+            async_flag: Some(Span::dummy("X")),
         };
         let result = block_on(item.execute(&mut env));
         assert_eq!(result, Break(Divert::Interrupt(Some(ExitStatus::NOEXEC))));